+++ app/src/pages/ErrorLoginPage/ErrorLoginPage.tsx
@@ -3,13 +3,13 @@
 // SPDX-License-Identifier: EUPL-1.2
 import { Container as MuiContainer, CssBaseline, Stack, styled, ThemeProvider, useTheme } from '@mui/material';
 import { WarningIcon } from '@opentalk/common';
-import React, { useEffect, useMemo, useState } from 'react';
+import { EventTypeError, useAuth } from '@opentalk/react-redux-appauth';
+import React, { useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { ReactComponent as Logo } from '../../assets/images/logo.svg';
 import { createOpenTalkTheme } from '../../assets/themes/opentalk';
 import Error from '../../commonComponents/Error';
-import LoginButton from '../../components/LoginPopup/fragments/LoginButton';
 import { useAppDispatch } from '../../hooks';
 import { hangUp } from '../../store/commonActions';
 import LobbyTemplate from '../../templates/LobbyTemplate';
@@ -33,11 +33,13 @@ interface ErrorLoginProps {
   error: Error;
 }
 
+const ERROR_MESSAGE_DISPLAY_BEFORE_SIGNOUT = 2500;
+
 const ErrorLoginPage = ({ error }: ErrorLoginProps) => {
-  const [displayLogin, setDisplayLogin] = useState<boolean>(false);
   const { t } = useTranslation();
   const theme = useTheme();
   const dispatch = useAppDispatch();
+  const { signOut } = useAuth();
 
   useEffect(() => {
     dispatch(hangUp());
@@ -45,16 +47,23 @@ const ErrorLoginPage = ({ error }: ErrorLoginProps) => {
 
   const errorDetails = useMemo(() => {
     switch (error.name) {
-      case 'error_session_expired':
-        setDisplayLogin(true);
+      case EventTypeError.SessionExpired:
+        setTimeout(() => {
+          signOut();
+        }, ERROR_MESSAGE_DISPLAY_BEFORE_SIGNOUT);
         return {
           name: t('error-session-expired'),
           message: t('error-session-expired-message'),
         };
-      case 'error_oidc_configuration':
+      case EventTypeError.OidcConfiguration:
         return {
           name: t('error-oidc-configuration'),
           message: t('error-oidc-configuration-message'),
+        };
+      case EventTypeError.SystemCurrentlyUnavailable:
+        return {
+          name: t('error-system-currently-unavailable'),
+          message: t('error-system-currently-unavailable'),
         };
       default:
         console.error('unknown login error:', error);
@@ -72,7 +81,6 @@ const ErrorLoginPage = ({ error }: ErrorLoginProps) => {
           <Stack spacing={8} alignItems="center">
             <WarningIcon fill={theme.palette.common.white} sx={{ width: '6rem', height: '6rem' }} />
             {errorDetails && <Error title={errorDetails.name} description={errorDetails.message} />}
-            {displayLogin && <LoginButton />}
             <StyledLogoIcon />
           </Stack>
         </Container>